requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.3
gunicorn==20.0.4
heroku==0.1.4
idna==2.9
numpy==1.18.1
opencv-python==4.2.0.32
pandas==1.0.1
Pillow==7.0.0
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
scipy==1.4.1
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
azure